src/requirements.txt
streamlit
pandas
numpy
plotly
scikit-learn
scipy
seaborn
imbalanced-learn
umap-learn